Ensembl scanning
git-svn-id: http://svn.systemsbiology.net/repos/compbio/Vesteinn_scripts/parse_external_data/trunk@10260 5021c4a6-a05a-4b22-8634-7b27b77db61f

+++ ScanEnsembl/ScanEnsembl.sh
@@ -8,11 +8,17 @@
 
 export CODE_DIR=~/parse_external_data/ScanEnsembl  ## Code and auxiliary files reside here
 
-cp -p ~/data/ncbi/gene_info_simplified_mouse . ## Archive this locally
-cp -p ~/data/ncbi/gene2ensembl_simplified_mouse . ## ditto
+if [ ! -f gene_info_simplified_mouse ]; then
+    cp -p ~/data/ncbi/gene_info_simplified_mouse . ## Archive this locally
+fi
+if [ ! -f gene2ensembl_simplified_mouse ] ; then
+    cp -p ~/data/ncbi/gene2ensembl_simplified_mouse . ## Archive this locally
+fi
 
 ## generate map file where ensembl ID is unique
 $CODE_DIR/UniquifyMapFile.py > EnsemblGeneIDtoEntrezGeneID.tsv
+
+## Requires results_fasta.txt
 
 #---------------------------------------
 # Prepare mapping files and scanned file
@@ -20,7 +26,10 @@ $CODE_DIR/UniquifyMapFile.py > EnsemblGeneIDtoEntrezGeneID.tsv
 $CODE_DIR/generateFastaFiles.py results_fasta.txt gene_info_simplified_mouse EnsemblGeneIDtoEntrezGeneID.tsv filelist > fastaFileList
 sed 's/fasta/gff/g' fastaFileList  > gffFileList
 $CODE_DIR/generateFastaFiles.py results_fasta.txt gene_info_simplified_mouse EnsemblGeneIDtoEntrezGeneID.tsv headers > fastaHeaders
-mkdir allMouseUpstreamFASTA
+
+if [ ! -d allMouseUpstreamFASTA ] ; then
+    mkdir allMouseUpstreamFASTA
+fi
 cd allMouseUpstreamFASTA
 $CODE_DIR/generateFastaFiles.py ../results_fasta.txt ../gene_info_simplified_mouse ../EnsemblGeneIDtoEntrezGeneID.tsv files
 #Generate all.fasta, a single big fasta file
@@ -35,22 +44,29 @@ R --vanilla < $CODE_DIR/EntrezEnsemblMaps.R
 #---------------------------------------
 # Perform scan, once for each motif
 #---------------------------------------
-mkdir  scanPerMotif
+if [ ! -d scanPerMotif ] ; then
+    mkdir  scanPerMotif
+fi
 $CODE_DIR/scanForEachMotif.sh all.fasta $CODE_DIR/q999.recomputed.tsv >& scanForEachMotif.log 
 #renice +18 PID
 
 #---------------------------------------
 # Generate promoter specific gff files from the per motif scans
 #---------------------------------------
-mkdir allMouseUpstreamGFF
+if [ ! -d allMouseUpstreamGFF ] ; then
+    mkdir allMouseUpstreamGFF
+fi
 $CODE_DIR/perGeneFromPerMotif.sh >& perGeneFromPerMotif.logfile 
 
 #---------------------------------------
 # Create Parsed gff files
 # Extract reduced set of columns from full GFF and convert to relative, upstream coordinates
 #---------------------------------------
-mkdir allMouseUpstreamParsedGFF
-$CODE_DIR/parseWrapper.sh >& parseWrapper.logfile 
+
+if [ ! -d allMouseUpstreamParsedGFF ] ; then
+    mkdir allMouseUpstreamParsedGFF
+fi
+$CODE_DIR/parseWrapper.sh >& parseWrapper.logfile
 
 #-------------------------------------------------
 # Create R version of scans
